+++ crates/base/src/board/piece.rs
@@ -42,7 +42,7 @@ impl Piece {
 
 impl Piece {
     #[inline(always)]
-    pub fn get_midgame_value(&self) -> isize {
+    pub fn get_midgame_value(&self) -> i32 {
         unsafe {
             let value = MIDGAME_PIECE_VALUE.get_unchecked(self.index());
             *value
@@ -50,7 +50,7 @@ impl Piece {
     }
 
     #[inline(always)]
-    pub fn get_endgame_value(&self) -> isize {
+    pub fn get_endgame_value(&self) -> i32 {
         unsafe {
             let value = ENDGAME_PIECE_VALUE.get_unchecked(self.index());
             *value
@@ -58,7 +58,7 @@ impl Piece {
     }
 
     #[inline(always)]
-    pub fn get_gamephase_value(&self) -> isize {
+    pub fn get_gamephase_value(&self) -> i32 {
         unsafe {
             let increment = GAMEPHASE_INCREMENT.get_unchecked(self.index());
             *increment
