+++ crates/lookup/src/tables/pesto.rs
@@ -11,7 +11,7 @@ pub const FLIP: [usize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_PAWN_TABLE: [isize; 64] = [
+pub const MIDGAME_PAWN_TABLE: [i32; 64] = [
       0,   0,   0,   0,   0,   0,  0,   0,
      98, 134,  61,  95,  68, 126, 34, -11,
      -6,   7,  26,  31,  65,  56, 25, -20,
@@ -23,7 +23,7 @@ pub const MIDGAME_PAWN_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const ENDGAME_PAWN_TABLE: [isize; 64] = [
+pub const ENDGAME_PAWN_TABLE: [i32; 64] = [
       0,   0,   0,   0,   0,   0,   0,   0,
     178, 173, 158, 134, 147, 132, 165, 187,
      94, 100,  85,  67,  56,  53,  82,  84,
@@ -35,7 +35,7 @@ pub const ENDGAME_PAWN_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_KNIGHT_TABLE: [isize; 64] = [
+pub const MIDGAME_KNIGHT_TABLE: [i32; 64] = [
     -167, -89, -34, -49,  61, -97, -15, -107,
      -73, -41,  72,  36,  23,  62,   7,  -17,
      -47,  60,  37,  65,  84, 129,  73,   44,
@@ -47,7 +47,7 @@ pub const MIDGAME_KNIGHT_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const ENDGAME_KNIGHT_TABLE: [isize; 64] = [
+pub const ENDGAME_KNIGHT_TABLE: [i32; 64] = [
     -58, -38, -13, -28, -31, -27, -63, -99,
     -25,  -8, -25,  -2,  -9, -25, -24, -52,
     -24, -20,  10,   9,  -1,  -9, -19, -41,
@@ -59,7 +59,7 @@ pub const ENDGAME_KNIGHT_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_BISHOP_TABLE: [isize; 64] = [
+pub const MIDGAME_BISHOP_TABLE: [i32; 64] = [
     -29,   4, -82, -37, -25, -42,   7,  -8,
     -26,  16, -18, -13,  30,  59,  18, -47,
     -16,  37,  43,  40,  35,  50,  37,  -2,
@@ -71,7 +71,7 @@ pub const MIDGAME_BISHOP_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const ENDGAME_BISHOP_TABLE: [isize; 64] = [
+pub const ENDGAME_BISHOP_TABLE: [i32; 64] = [
     -14, -21, -11,  -8, -7,  -9, -17, -24,
      -8,  -4,   7, -12, -3, -13,  -4, -14,
       2,  -8,   0,  -1, -2,   6,   0,   4,
@@ -83,7 +83,7 @@ pub const ENDGAME_BISHOP_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_ROOK_TABLE: [isize; 64] = [
+pub const MIDGAME_ROOK_TABLE: [i32; 64] = [
      32,  42,  32,  51, 63,  9,  31,  43,
      27,  32,  58,  62, 80, 67,  26,  44,
      -5,  19,  26,  36, 17, 45,  61,  16,
@@ -95,7 +95,7 @@ pub const MIDGAME_ROOK_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const ENDGAME_ROOK_TABLE: [isize; 64] = [
+pub const ENDGAME_ROOK_TABLE: [i32; 64] = [
     13, 10, 18, 15, 12,  12,   8,   5,
     11, 13, 13, 11, -3,   3,   8,   3,
      7,  7,  7,  5,  4,  -3,  -5,  -3,
@@ -107,7 +107,7 @@ pub const ENDGAME_ROOK_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_QUEEN_TABLE: [isize; 64] = [
+pub const MIDGAME_QUEEN_TABLE: [i32; 64] = [
     -28,   0,  29,  12,  59,  44,  43,  45,
     -24, -39,  -5,   1, -16,  57,  28,  54,
     -13, -17,   7,   8,  29,  56,  47,  57,
@@ -119,7 +119,7 @@ pub const MIDGAME_QUEEN_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const ENDGAME_QUEEN_TABLE: [isize; 64] = [
+pub const ENDGAME_QUEEN_TABLE: [i32; 64] = [
      -9,  22,  22,  27,  27,  19,  10,  20,
     -17,  20,  32,  41,  58,  25,  30,   0,
     -20,   6,   9,  49,  47,  35,  19,   9,
@@ -131,7 +131,7 @@ pub const ENDGAME_QUEEN_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_KING_TABLE: [isize; 64] = [
+pub const MIDGAME_KING_TABLE: [i32; 64] = [
     -65,  23,  16, -15, -56, -34,   2,  13,
      29,  -1, -20,  -7,  -8,  -4, -38, -29,
      -9,  24,   2, -16, -20,   6,  22, -22,
@@ -143,7 +143,7 @@ pub const MIDGAME_KING_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const ENDGAME_KING_TABLE: [isize; 64] = [
+pub const ENDGAME_KING_TABLE: [i32; 64] = [
     -74, -35, -18, -18, -11,  15,   4, -17,
     -12,  17,  14,  17,  17,  38,  23,  11,
      10,  17,  23,  15,  20,  45,  44,  13,
@@ -155,10 +155,10 @@ pub const ENDGAME_KING_TABLE: [isize; 64] = [
 ];
 
 #[rustfmt::skip]
-pub const MIDGAME_PIECE_VALUE: [isize; 7] = [0, 82, 337, 365, 477, 1025, 0];
+pub const MIDGAME_PIECE_VALUE: [i32; 7] = [0, 82, 337, 365, 477, 1025, 0];
 
 #[rustfmt::skip]
-pub const ENDGAME_PIECE_VALUE: [isize; 7] = [0, 94, 281, 297, 512, 936, 0];
+pub const ENDGAME_PIECE_VALUE: [i32; 7] = [0, 94, 281, 297, 512, 936, 0];
 
 #[rustfmt::skip]
-pub const GAMEPHASE_INCREMENT: [isize; 7] = [0, 0, 1, 1, 2, 4, 0];
+pub const GAMEPHASE_INCREMENT: [i32; 7] = [0, 0, 1, 1, 2, 4, 0];
